+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/ins1d30ut/apps/max-zip/01-starting-setup/src/index.js":"1","/home/ins1d30ut/apps/max-zip/01-starting-setup/src/App.js":"2","/home/ins1d30ut/apps/max-zip/01-starting-setup/src/components/ExpenseItem.js":"3","/home/ins1d30ut/apps/max-zip/01-starting-setup/src/components/ExpenseDate.js":"4"},{"size":179,"mtime":1648532578000,"results":"5","hashOfConfig":"6"},{"size":736,"mtime":1664656124879,"results":"7","hashOfConfig":"6"},{"size":394,"mtime":1664655576647,"results":"8","hashOfConfig":"6"},{"size":500,"mtime":1664655683729,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"og6fqd",{"filePath":"12","messages":"13","errorCount":12,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/ins1d30ut/apps/max-zip/01-starting-setup/src/index.js",[],"/home/ins1d30ut/apps/max-zip/01-starting-setup/src/App.js",["18","19","20","21","22","23","24","25","26","27","28","29"],"/home/ins1d30ut/apps/max-zip/01-starting-setup/src/components/ExpenseItem.js",[],"/home/ins1d30ut/apps/max-zip/01-starting-setup/src/components/ExpenseDate.js",[],{"ruleId":"30","severity":2,"message":"31","line":8,"column":16,"nodeType":"32","messageId":"33","endLine":8,"endColumn":24},{"ruleId":"30","severity":2,"message":"31","line":9,"column":17,"nodeType":"32","messageId":"33","endLine":9,"endColumn":25},{"ruleId":"30","severity":2,"message":"31","line":10,"column":15,"nodeType":"32","messageId":"33","endLine":10,"endColumn":23},{"ruleId":"30","severity":2,"message":"31","line":13,"column":16,"nodeType":"32","messageId":"33","endLine":13,"endColumn":24},{"ruleId":"30","severity":2,"message":"31","line":14,"column":17,"nodeType":"32","messageId":"33","endLine":14,"endColumn":25},{"ruleId":"30","severity":2,"message":"31","line":15,"column":15,"nodeType":"32","messageId":"33","endLine":15,"endColumn":23},{"ruleId":"30","severity":2,"message":"31","line":18,"column":16,"nodeType":"32","messageId":"33","endLine":18,"endColumn":24},{"ruleId":"30","severity":2,"message":"31","line":19,"column":17,"nodeType":"32","messageId":"33","endLine":19,"endColumn":25},{"ruleId":"30","severity":2,"message":"31","line":20,"column":15,"nodeType":"32","messageId":"33","endLine":20,"endColumn":23},{"ruleId":"30","severity":2,"message":"31","line":23,"column":16,"nodeType":"32","messageId":"33","endLine":23,"endColumn":24},{"ruleId":"30","severity":2,"message":"31","line":24,"column":17,"nodeType":"32","messageId":"33","endLine":24,"endColumn":25},{"ruleId":"30","severity":2,"message":"31","line":25,"column":15,"nodeType":"32","messageId":"33","endLine":25,"endColumn":23},"no-undef","'expenses' is not defined.","Identifier","undef"]
+[{"/home/ins1d30ut/apps/max-zip/react-comp/src/index.js":"1","/home/ins1d30ut/apps/max-zip/react-comp/src/App.js":"2","/home/ins1d30ut/apps/max-zip/react-comp/src/components/ExpenseItem.js":"3","/home/ins1d30ut/apps/max-zip/react-comp/src/components/ExpenseDate.js":"4","/home/ins1d30ut/apps/max-zip/react-comp/src/components/Expenses.js":"5"},{"size":179,"mtime":1664657708693,"results":"6","hashOfConfig":"7"},{"size":648,"mtime":1664657681577,"results":"8","hashOfConfig":"7"},{"size":394,"mtime":1664657627433,"results":"9","hashOfConfig":"7"},{"size":500,"mtime":1664655683729,"results":"10","hashOfConfig":"7"},{"size":736,"mtime":1664657599904,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1tb596i",{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/ins1d30ut/apps/max-zip/react-comp/src/index.js",[],"/home/ins1d30ut/apps/max-zip/react-comp/src/App.js",[],"/home/ins1d30ut/apps/max-zip/react-comp/src/components/ExpenseItem.js",[],"/home/ins1d30ut/apps/max-zip/react-comp/src/components/ExpenseDate.js",[],"/home/ins1d30ut/apps/max-zip/react-comp/src/components/Expenses.js",[]]
